refactor(crane-context): split queries.ts into per-handler submodules
Brings workers/crane-context/src/endpoints/queries.ts under structural caps
(was 1079 LOC).

- queries.ts (orchestrator, 303 LOC): retains simpler handlers
  (handleGetActiveSessions, handleUpdateHandoffStatus, handleListDocsPublic,
  handleGetVentures, handleDocAudit, handleGetDoc) + re-exports from submodules.
- queries/session-history.ts: handleGetSessionHistory + grouping/activity
  helpers (buildActivityRanges, mergeSessionsIntoBlocks).
- queries/prior-session.ts: handleGetPriorSession + param parsing.
- queries/handoff-latest.ts: handleGetLatestHandoff + filter parsing.
- queries/handoff-query.ts: handleQueryHandoffs + mode/date/limit parsing.

Public export surface unchanged. 540/540 tests pass.
